+++ src/main/java/chess/domain/piece/Pawn.java
@@ -1,9 +1,7 @@
-package chess.domain.piece.unitmovepiece;
+package chess.domain.piece;
 
 import java.util.List;
 
-import chess.domain.piece.Color;
-import chess.domain.piece.Piece;
 import chess.domain.position.Direction;
 import chess.domain.position.UnitDirection;
