+++ src/test/java/chess/domain/piece/multiplemovepiece/RookTest.java
@@ -6,9 +6,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import chess.domain.piece.Color;
+import chess.domain.piece.None;
 import chess.domain.piece.mulitiplemovepiece.Bishop;
 import chess.domain.piece.mulitiplemovepiece.Rook;
-import chess.domain.piece.unitmovepiece.None;
 import chess.domain.position.Direction;
 
 public class RookTest {
